refactor(ConfigService): extract mergeDeviceNames as pure utility
Fixes a real prod/test drift bug. The deviceNames unit test was
maintaining a copy-pasted version of mergeDeviceNames that had diverged
from the production implementation: the test used array spread for
tombstone aggregation, while production used [...new Set([...])] for
deduplication. The test could pass while the prod code silently
de-duplicated repeated tombstone entries.

Changes:
- New module src/services/configMergeHelpers.ts holds the pure
  mergeDeviceNames function (additive union + dedup'd tombstones).
  Mirrors the existing channelThreadHelpers.ts pattern.
- ConfigService imports and calls the helper instead of the private
  method (which is removed).
- The unit test imports the real function instead of a local copy.
- Added a regression test pinning the dedup behaviour:
  mergeDeviceNames({}, {}, ['QmAAA'], ['QmAAA'])
    -> deletedDeviceNameAddresses === ['QmAAA']  // not ['QmAAA', 'QmAAA']
  This would have caught the prior drift had it existed.
- Added a remote-only-tombstone test (remote provides a name and also
  tombstones it -> name removed from result).

Verified:
- 24 test files, 455 tests + 2 todo, all green (+2 tests from new dedup
  + remote-tombstone cases)
- npx tsc --noEmit --skipLibCheck clean (one pre-existing unrelated
  error in ImportKeyStep.tsx not touched by this change)

The function is pure and now naturally migration-ready if/when
ConfigService's privacy-fields-consolidation work lifts it into
@quilibrium/quorum-shared.
